requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
user-agents==2.2.0
geoip2==4.7.0
pillow==10.2.0
python-magic==0.4.27
aiofiles==23.2.1 